Folding for functions with parameters over multiple lines
The default configuration fails to fold functions if the parameters
are defined over multiple lines. Due to typing, this is more likely
to happen these days than it used to be.

Three things here:

- a new regex for foldingIndentedBlockStart.

The existing expression only matches when the final : is on the same
line as the 'def'. We don't want to lose that requirement for the other
keywords that start a block, so I added a totally separate expression
that matches only functions definitions with a open-parenthesis
as the final non-whitespace non-comment character on the line.

`^\s*def\b.*\(\s*(#.*)?$`

- adding a foldingIndentedBlockIgnore expression

This is needed so that the closing parenthesis of the parameter
declaration can be on a lower indent level, like black likes to do.
It should match a closing parenthesis and colon with optional return
type.

`^\s*\)(\s*->\s*\w.*)?:\s*(#.*)?$`

- modifying foldingStartMarker

Still not done, because Textmate also matches `def foo(\n` as
the start of marker-based folding due to the parenthesis, and
then it gets confused. I added a negative lookbehind to prevent
it from matching when we're making a def statement.

`(?<!\bdef\s+\w+\s*)`

With those three changes in place, Textmate can now fold
all functions, not just those with `def` and `):` on the same line!